Cleaned up POS.clean()
Removed per field checks of existence and non-blank, this is dealt with
automatically through the blank argument to the field constructor (which
defaults to True).
Made the check that status is reinforced in order to add an rftime raise a
ValidationError rather than silently discard the provided rftime as throwing
away input without telling anyone tends to piss people off.