Factored out trusted igb functionality in map_checkin
Moved all of the functionality using a trusted igb (automatic adding of systems
when we detect a jump to a system we don't know about) into a seperate
function.
Found and fixed a bug where it would try to add systems for all k-space ->
k-space jumps so we'd end up with everybody's routes out to jita cluttering up
the map. There is still a slightly unresolved issue with k-space -> w-space
jumps being but I think it should be fine to rely on users to decide whether to
add those when prompted
Started moving functionality out of utils.py and into relevent Model
declarations as object methods.
Some additional slight cleanup as I went.